Split dialects to seperate files.
